PathPrimites extenced
PathPrimitives now ads random curves. The Path is still in no way
meaningful. It still remains to be decided what primitives should be
possible in how much of the path control should be given to this
"algorithm" and how much impossibility simply has to be solved by the
path following.