GenomeToPath
Lots of bugfixes and comments.
There is now a function that creates a Path from a given Genome which is
needed for the
fitness function after mutating/crossing genomes.
PathToGenome does not seem necessary right now since a Path is either
randomly generated (which also generates the genome) or created by
mutating/crossing other genomes in which case the genome is already
given.