Debug Code
The Error occurs when the angle is negativ (which makes sense since i
shouldnt be negative).
It is negative whenever startangle (for right turns) or endangle (for
left turns) is negative, which also makes sense since those also
shouldnt be negative. The error most likely occurs in the AngleWrapper
function which returns negative values even though it shouldnt.